Adding a few cities that I visited in 2019
I spent two weeks on the Spanish islands of Ibiza and Formentera over the
summer.

I went to Hamburg twice to visit Mimi's family - once in the summer; once for
Christmas.

In the Fall, I went to Bordeaux with Mimi where we stayed at a charming Airbnb.

I spent New Years Eve in Amsterdam with Matty, Ryan, and Conor.

I may be missing a few other places that I visited in 2019; it was an active
year.