chore(web/tvl): Remove people I don't know directly
Quoting myself from IRC, for those who missed it:

-------------

Alright, some of you might be wondering what 'UNDERGOING CHANGES'
means. The gist of it is that TVL has kind of departed from what it
was originally (a place for friends of mine to hang out) by growing a
little too fast, and I've decided to do a bit of a reboot.

What this means is that for most people I don't know directly, I'll be
asking you to leave (I'll +v/-v accordingly) and see if we can get the
original TVL crowd back before opening it for a wider audience again.
The "opening hours" (heh) will also be more restricted again.

Some people will be a little unhappy about this, but the good news is
that you can easily make your own Meet and use that! Some folks even
started an alternative EU-daytime lounge already. But for now, with
this particular one, it'll be s/The V/tazjin's V/.

----------------