Update patch set 1
Patch Set 1:

Does this actually work out-of-the-box? My expectation is that Gerrit will probably attempt to template something into that URL, but the cgit is patched to run the depot at the root.

I think this might require a URL rewrite in nginx, because I'd be surprised if Gerrit lets us configure this the right way.

Patch-set: 1
Reviewer: Gerrit User 1000001 <1000001@4fdfa107-4df9-4596-8e0a-1d2bbdd96e36>