Update patch set 1
Patch Set 1:

> Should we have some tests for some of these?

I'd like that, but it's not currently supported (other than testing them interactively while writing them). The thing is that we can't easily run them in pure-Prolog because we would need to pull out all the Prolog files from our dependencies (like the owners plugin), which would be annoying.

It'd be interesting to investigate the options we have (including writing a Gerrit plugin).

Patch-set: 1