feat: Configure OWNERS file for the depot root
This owners file creates a default approval policy for the entire
tree.

This is just an initial version and we haven't actually decided what
this should look like, but crucially this change means that we can
allow everyone to +2 changes in the repository which allows everyone
to reign freely within their user directories.

Change-Id: I2a14092765ad484d329f5542ac052da59eb104c7
Reviewed-on: https://cl.tvl.fyi/c/depot/+/259
Reviewed-by: lukegb <lukegb@tvl.fyi>