feat: Configure depot access for git-bug
git-bug uses two different ref types (as far as I can tell) for its bug tracking:

* refs/bugs/* - this is where the bugs go
* refs/identities/* - this is where the bug author identities go

For the former, every user should be allowed to push (but not force-push). This way any change made is append-only due to the structure of git-bug.

For the latter, reviews should be required. I'm not entirely sure if this is configured correctly via this commit, but we'll find out.

Change-Id: Icd5a28e986b21c1c15b8708e7d8e9705b865b922
Reviewed-on: https://cl.tvl.fyi/c/depot/+/362
Reviewed-by: q3k <q3k@q3k.org>