refactor(nix/readTree): Use 'passthru' to merge values into drvs
readTree merges the `__readTree` attribute as well as any children in
the directory structure into an attribute set.

This leads to situations where derivations contain other derivations,
even though they're not directly part of the builder.

This moves all merging into the `passthru` value where applicable,
which means that they will be ignored by the builder.

Since this is a feature of stdenv and not of the derivation itself, a
check is performed to see whether the set is a stdenv derivation.

Change-Id: Ieeb80281cfcbc797e243b3bc6dcb64a67722f991