feat(rules): Replace default rules with explicit checks
Instead of using the default checks, we now specify the Code-Review
check that is required explicitly.

This by extension means that the Verified check is not required
anymore, which we unfortunately introduced when making that label
available by default.

The Verified check will be added back when the CI build server is
permanently online.

Since we don't need to unpack the default checks anymore the code is
now also somewhat simplified (in particular, append is no longer
required).

Change-Id: I9e624cb04b0a434fc5b69228b25f21c25d76e0d3
Reviewed-on: https://cl.tvl.fyi/c/depot/+/707
Tested-by: tazjin <mail@tazj.in>
Reviewed-by: Kane York <rikingcoding@gmail.com>