Update patch set 1
Abandoned

This is no longer necessary, because we'll need to make a Buildkite hook instead (and those variables by extension don't need to be passed to Buildkite anymore).

I will make an amended version of this CL that creates a link back to Gerrit in the message on Buildkite, though.

Patch-set: 1
Status: abandoned
Tag: autogenerated:gerrit:abandon