chore(cheddar): Vendor the HTML writer implementation into cheddar
We want to modify the HTML renderer used by cheddar to support some
custom features. To make this possible, this commit vendors the code
into cheddar.

This now contains a whole bunch of code that is not actually used by
cheddar, but I wanted to leave a reasonable trail of commits for this.

Change-Id: I9102dfd09720d2272430c9105bc9d4452b388db8