feat(ops/nixos): Add generic install-system script
This adds a first crack at one idea for a generic, non-user-specific
install-system script to ops.nixos.install-system. The idea here is that
we enumerate all the nixos systems stored in the monorepo (similarly to
what we do for ci-builds right now) then search through them by hostname
to find the one matching the hostname of the current system, which is an
attempt at a more generic version of tazjin's rebuilder script which
does the same thing but with an explicit case block.

Some caveats I see with this approach:

- To get around the fact that different people might have the depot
  cloned in different places, and not everyone is going to have it
  exported on their NIX_PATH, this uses a horrifically impure `toString
  ../..` to embed the clone path of the depot inside of the built shell
  script. This feels fine to *me* but if other people feel uncomfortable
  with it I could be convinced that just expecting people to have the
  depot on their `NIX_PATH` is the way to go
- There's a slight possibility that this way of finding systems is going
  to get slow to evaluate - on my system it feels fine but if it grows
  out of hand it's probably feasible to just bake this into the built
  script as a dynamically generated case statement.

Change-Id: I2e4c5401913b6f4d936ab48ba2f95f96e0e78eb4