Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> Isn't this something that should be done in a NixOS module instead? It seems weird to include this configuration in the package build.

In practice, the _entire package build_ is the configuration, because you have to tweak the build.gradle and such in order to get this to do the right thing, and you're expected to overlay a bunch of other crap directly into this repo. I decided it was easier to just completely special-case the package rather than trying to make it generic.

Patch-set: 2