Update patch set 1
Patch Set 1:

> Patch Set 1: Code-Review+1
> 
> wonder how easy this would be to unit test...
> 
> as a non-c++ person, that array index terrifies me

Well, Nix has no unit tests so it would first involve setting up the entire test scaffolding. That's on the todo list, but not right now.

This array index is safe because the condition check happens immediately before it and this isn't being accessed concurrently (I ... think, if it is then we have other problems anyways).

Patch-set: 1