Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1: Code-Review+1
> > 
> > wonder how easy this would be to unit test...
> > 
> > as a non-c++ person, that array index terrifies me
> 
> Well, Nix has no unit tests so it would first involve setting up the entire test scaffolding. That's on the todo list, but not right now.

There are two tests in the `tests` folder that are currently spawning a daemon. I am working on a CS locally to see if that is as easy as I imagine. Generally I've been involved in adding a few unit tests to Nix upstream using the Google Test Framework. I think most of them should be portable to this repo if we desired to.
 
> This array index is safe because the condition check happens immediately before it and this isn't being accessed concurrently (I ... think, if it is then we have other problems anyways).

Yeah, there are other things in the Nix code base where the string access is basically UB but in practice (due to SSO) do not yield to any ASAN warning or "real" issues. I wish they did. That just encourages sloppy code.

Patch-set: 1
CC: Gerrit User 1000017 <1000017@4fdfa107-4df9-4596-8e0a-1d2bbdd96e36>