feat(3p/nix): Add vector-backed impl for Bindings
Add an alternative impl of the now-abstract Bindings base class that is
backed by a std::vector, somewhat similar but stylistically a little
superior to the array-backed implementation in upstream nix. The
underlying iterator type in BindingsIterator is now backed by a
std::variant that we std::visit an overload over in order to implement
the various bits of the iterator interface.

Change-Id: I7fbd1f4d5c449e2f9b82102a701b0bacd5e80672